Add a few extra code to test/count the number of iteration to reach the fix-point.
A conjecture is that, the max numb. of steps is
(one step correspond to one stream of the fixpoint function):

local x1 do local x2 do ... local xk in eq
with n equations in eq is bounded by (2^k).n
with the very naive iteration algorithm implemented in
zrun. This would meen that the complexity is exponential
in the deepth of function nodes, at worst. In actual programs,
this is low (I meen, uncomparably less than the size of the program).