Bug fix for the Lustre interpretation of the if/then/else.
if bot then _ else _ = bot
if nil then _ else _ = nil

if _ then bot else _ = bot
if _ then _ else bot = bot

but:

if true then v else nil = v
if false then nil else v = v

whereas what was implemented was returning bot (or nil) as soon as one is bot (or nil). The
consequence is that, with the Lustre interpretation:

x = 0 -> pre x + 1

was producing an infinite sequence of nil !

This is fixed.