Evaluation of const/static declarations.
[let const E in e] and [let static E in e]
are evaluated during the initialization.

No difference for the moment.

To be done:

- redefine the type of environment to distinguish
const/static values from dynamic ones.

- decide what to do between const and static. The
initial intend was that "const" is prescriptive;
it must be done during compilation; "static"
is done during the instantiation, that is,
the initial code that is computed before
the one that is cyclically executed.