Bug fix for loops with resume.
- still, the semantics is overly complicated; in
particular, I am not sure we need the two
functions set_vardec and set_for_out.

... returns (x init e) ...

- The idea was that variables from [e] are
not the one from the body. That is, the
scope of e is the global one. 

- [e] is evaluated at every step.
- the update of [last x] that is stored in
the state can only be done one we know x.

- The current implementation considers that
accumulated values are stored in [last x] in
the environment produced by the body of the loop.
This is not good. We should store it in [x] and
also keep [last x].