Update
Not finished.

- mutually recursive definitions
of size functions does not work.

- when calling f<42> where:

f = Vsizefix { bound; f;
                      f1 = ...; ...; fn = ...}

the corresponding function associated to f must be evaluated but, because f is recursive, one must record that the new value for the bound in 42 in any further use of f1,..., fn. This is not done with the existing encoding.

Adding an extra parameter to
the value of f, i.e.,
f : ientry Ent.t -> int list -> value result ?