First version - it compiles !
- Values are now more "denotational"
(the value of a function definition
is a function, instead of the source
code that is further interpreted).
- This is more conventional w.r.t classical denotational semantics.
- This means that the use of the
semantics for the code specialisation
done in Zelus will have to be changed deeply. A good way would be to make Zelus more generic, with two possible implementations for values
(one implicit - denotational - one explicit/symbolic - where the
value of a function is its text).